fix(demo): Timeline, adding ids, reorganisation of some knobs - FRONT-972
