+++ src/ec/packages/ec-component-timeline/README.md
@@ -11,8 +11,8 @@ npm install --save @ecl-twig/ec-component-timeline
 - **"toggle_collapsed"** (string) (default: '')
 - **"toggle_expanded"** (string) (default: '')
 - **"hide"** (object): (default: undefined)
-  - "from": (integer) (default: undefined) Item index after which to start hiding timeline items
-  - "to": (integer) (default: undefined) Item index after which to resume displaying timeline items
+  - "from": (integer) (default: items.length) Item index after which to start hiding timeline items
+  - "to": (integer) (default: items.length) Item index after which to resume displaying timeline items
 - **"items"** (array) (default: []):
   - "id": (string) (default: '')
   - "label": (string) (default: '')
@@ -28,7 +28,7 @@ npm install --save @ecl-twig/ec-component-timeline
 
 <!-- prettier-ignore -->
 ```twig
-{% include '@ecl-twig/ec-component-timeline/ecl-timeline.html.twig' with {  
+{% include '@ecl-twig/ec-component-timeline/ecl-timeline.html.twig' with { 
   toggle_collapsed: 'Show 10 more items', 
   toggle_expanded: 'Hide 10 items', 
   hide: {from: 7, to: -2},
