+++ src/ec/packages/ec-component-timeline/ecl-timeline.html.twig
@@ -5,8 +5,8 @@ Parameters:
   - "toggle_collapsed": (string) (default: '')
   - "toggle_expanded": (string) (default: '')
   - "hide" (object): (default: undefined)
-    - "from": (integer) (default: undefined) Item index after which to start hiding timeline items
-    - "to": (integer) (default: undefined) Item index after which to resume displaying timeline items
+    - "from": (integer) (default: items.length) Item index after which to start hiding timeline items
+    - "to": (integer) (default: items.length) Item index after which to resume displaying timeline items
   - "items" (array) (default: []):
     - "id": (string) (default: '')
     - "label": (string) (default: '')
