+++ src/ec/packages/ec-component-timeline/timeline.story.js
@@ -3,12 +3,11 @@ import { storiesOf } from '@storybook/html';
 import { withKnobs, number, text, select } from '@storybook/addon-knobs';
 import { withNotes } from '@ecl-twig/storybook-addon-notes';
 import { getExtraKnobs, buttonLabels } from '@ecl-twig/story-utils';
-import withCode from '@ecl-twig/storybook-addon-code';
-import iconPath from '@ecl/ec-resources-icons/dist/sprites/icons.svg';
 import he from 'he';
+import withCode from '@ecl-twig/storybook-addon-code';
 
+import iconPath from '@ecl/ec-resources-icons/dist/sprites/icons.svg';
 import demoData from './demo/data';
-
 import timeline from './ecl-timeline.html.twig';
 import notes from './README.md';
 
@@ -21,18 +20,6 @@ const prepareTimeline = data => {
   } else {
     hiddenCount = data.items.length + to - from;
   }
-  data.toggle_collapsed = select(
-    'data.toggle_collapsed',
-    [`Show ${hiddenCount} more items`],
-    `Show ${hiddenCount} more items`,
-    buttonLabels.required
-  );
-  data.toggle_expanded = select(
-    'data.toggle_expanded',
-    [`Hide ${hiddenCount} items`],
-    `Hide ${hiddenCount} items`,
-    buttonLabels.required
-  );
   data.hide.from = number(
     'hide.from',
     data.hide.from,
@@ -40,21 +27,35 @@ const prepareTimeline = data => {
     buttonLabels.required
   );
   data.hide.to = number('hide.to', data.hide.to, {}, buttonLabels.required);
-  data.items.forEach((item, index) => {
-    const { id, label, title, content } = item;
-    item.id = text(`items[${index}].id`, id, buttonLabels.required);
-    item.label = text(`items[${index}].label`, label, buttonLabels.required);
-    item.title = text(`items[${index}].title`, title, buttonLabels.required);
-    item.content = he.decode(
-      text(`items[${index}].content`, content, buttonLabels.required)
-    );
-  });
+  data.toggle_collapsed = select(
+    'toggle_collapsed',
+    [`Show ${hiddenCount} more items`],
+    `Show ${hiddenCount} more items`,
+    buttonLabels.required
+  );
+  data.toggle_expanded = select(
+    'toggle_expanded',
+    [`Hide ${hiddenCount} items`],
+    `Hide ${hiddenCount} items`,
+    buttonLabels.required
+  );
   data.icon_path = select(
     'icon_path',
     [iconPath],
     iconPath,
     buttonLabels.required
   );
+
+  data.items.forEach((item, index) => {
+    const { id, label, title, content } = item;
+    item.id = text(`items[${index}].id`, id, buttonLabels.optional);
+    item.label = text(`items[${index}].label`, label, buttonLabels.required);
+    item.title = text(`items[${index}].title`, title, buttonLabels.optional);
+    item.content = he.decode(
+      text(`items[${index}].content`, content, buttonLabels.required)
+    );
+  });
+
   getExtraKnobs(data);
 
   return data;
